refactor: migrate shared hooks to shared/hooks with stubs
