+++ packages/commonwealth/client/scripts/hooks/useAppStatus.ts
@@ -1,20 +1,2 @@
-const useAppStatus = () => {
-  const isAddedToHomeScreen = window.matchMedia(
-    '(display-mode: standalone)',
-  ).matches;
-  const isMarketingPage = window.location.pathname === '/';
-  const isIOS =
-    !!window.navigator.userAgent.match(/(iPad|iPhone|iPod)/g) ||
-    (navigator.platform === 'MacIntel' && navigator.maxTouchPoints > 1);
-
-  const isAndroid = window.navigator.userAgent.match(/Android/g) ? true : false;
-
-  return {
-    isAddedToHomeScreen,
-    isMarketingPage,
-    isIOS,
-    isAndroid,
-  };
-};
-
-export default useAppStatus;
+export * from 'shared/hooks/useAppStatus';
+export { default } from 'shared/hooks/useAppStatus';
