+++ packages/commonwealth/client/scripts/hooks/useBeforeUnload.ts
@@ -1,20 +1,2 @@
-import { useEffect } from 'react';
-
-const useBeforeUnload = (enabled = true) => {
-  useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      if (enabled) {
-        event.preventDefault();
-        event.returnValue = true;
-      }
-    };
-
-    window.addEventListener('beforeunload', handleBeforeUnload);
-
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-    };
-  }, [enabled]);
-};
-
-export default useBeforeUnload;
+export * from 'shared/hooks/useBeforeUnload';
+export { default } from 'shared/hooks/useBeforeUnload';
