+++ packages/commonwealth/client/scripts/hooks/useFlag.ts
@@ -1,10 +1,1 @@
-import { useBooleanFlagValue } from '@openfeature/react-sdk';
-import { AvailableFeatureFlag } from '../helpers/feature-flags';
-
-export const useFlag = (name: AvailableFeatureFlag): boolean => {
-  return (
-    useBooleanFlagValue(name, false, {
-      updateOnConfigurationChanged: false,
-    }) ?? false
-  );
-};
+export * from 'shared/hooks/useFlag';
