+++ packages/commonwealth/client/scripts/shared/hooks/useBeforeUnload.ts
@@ -1,0 +1,20 @@
+import { useEffect } from 'react';
+
+const useBeforeUnload = (enabled = true) => {
+  useEffect(() => {
+    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
+      if (enabled) {
+        event.preventDefault();
+        event.returnValue = true;
+      }
+    };
+
+    window.addEventListener('beforeunload', handleBeforeUnload);
+
+    return () => {
+      window.removeEventListener('beforeunload', handleBeforeUnload);
+    };
+  }, [enabled]);
+};
+
+export default useBeforeUnload;
