+++ packages/commonwealth/client/scripts/shared/hooks/useFlag.ts
@@ -1,0 +1,10 @@
+import { useBooleanFlagValue } from '@openfeature/react-sdk';
+import { AvailableFeatureFlag } from 'helpers/feature-flags';
+
+export const useFlag = (name: AvailableFeatureFlag): boolean => {
+  return (
+    useBooleanFlagValue(name, false, {
+      updateOnConfigurationChanged: false,
+    }) ?? false
+  );
+};
