+++ packages/commonwealth/client/scripts/shared/hooks/useForceRerender.ts
@@ -1,0 +1,13 @@
+import { useCallback, useState } from 'react';
+
+const useForceRerender = () => {
+  const [, setState] = useState({});
+
+  const forceRerender = useCallback(() => {
+    setState({});
+  }, []);
+
+  return forceRerender;
+};
+
+export default useForceRerender;
